+++ pipelex/config.py
@@ -46,6 +46,16 @@ class PromptingConfig(ConfigModel):
 class FeatureConfig(ConfigModel):
     is_pipeline_tracking_enabled: bool
     is_activity_tracking_enabled: bool
+    is_reporting_enabled: bool
+
+
+class ReportingConfig(ConfigModel):
+    is_log_costs_to_console: bool
+    is_generate_cost_report_file_enabled: bool
+    cost_report_dir_path: str
+    cost_report_base_name: str
+    cost_report_extension: str
+    cost_report_unit_scale: float
 
 
 class Pipelex(ConfigModel):
@@ -61,6 +71,7 @@ class Pipelex(ConfigModel):
     prompting_config: PromptingConfig
 
     pipe_run_config: PipeRunConfig
+    reporting_config: ReportingConfig
 
 
 class PipelexConfig(ConfigRoot):